backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
llama-index-core
llama-index-readers-file
llama-index-llms-openai-like
python-dotenv
pydantic>=2.0
